refactor : 회원 가입, 로그인, 멤버 기능 분리
