+++ src/main/java/miniProject/board/controller/member/MemberController.java
@@ -1,0 +1,14 @@
+package miniProject.board.controller.member;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.*;
+
+@Slf4j
+@Controller
+@RequestMapping("/members")
+@RequiredArgsConstructor
+public class MemberController {
+
+}
