+++ src/main/java/miniProject/board/service/member/MemberService.java
@@ -1,14 +1,10 @@
-package miniProject.board.service;
+package miniProject.board.service.member;
 
 import miniProject.board.dto.MemberDto;
 
 import java.util.List;
 
 public interface MemberService {
-    MemberDto.Info signUp(MemberDto.Create createMemberDto);
-
-    MemberDto.Session login(MemberDto.Login loginMemberDto);
-
     MemberDto.Info findMember(Long memberId);
 
     List<MemberDto.Info> findMembers();
